게시판과 user 을 묶었습니다.

+++ src/main/java/com/study/boardPage/board/application/BoardService.java
@@ -8,6 +8,9 @@ import com.study.boardPage.board.dto.resp.BoardReadDto;
 import com.study.boardPage.board.infrastructure.BoardRepository;
 import com.study.boardPage.global.exception.BaseException;
 import com.study.boardPage.global.response.ErrorCode;
+import com.study.boardPage.security.JwtHTokenHeaderFilter;
+import com.study.boardPage.users.domain.Users;
+import com.study.boardPage.users.infrastructure.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
@@ -18,9 +21,13 @@ import java.util.stream.Collectors;
 @Service
 public class BoardService {
     private final BoardRepository boardRepository;
+    private final JwtHTokenHeaderFilter jwtHTokenHeaderFilter;
+    private final UserRepository userRepository;
 
-    public BoardService(BoardRepository boardRepository) {
+    public BoardService(BoardRepository boardRepository , JwtHTokenHeaderFilter jwtHTokenHeaderFilter, UserRepository userRepository) {
         this.boardRepository = boardRepository;
+        this.jwtHTokenHeaderFilter = jwtHTokenHeaderFilter;
+        this.userRepository = userRepository;
     }
     public BoardReadDto getBoardById(int id) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new BaseException(ErrorCode.BOARD_READ_FAILED));
@@ -44,29 +51,44 @@ public class BoardService {
         return new BoardAllDto(boardList);
     }
 
-    public Integer addBoard(BoardCreateDto boardCreateDto) {
+    public Integer addBoard(BoardCreateDto boardCreateDto, String token ) {
         if (boardCreateDto == null) {
             throw new BaseException(ErrorCode.BOARD_CREATE_FAILED);
         }
+
+        String userId = jwtHTokenHeaderFilter.getUserIdFromToken(token);
+        Users users = userRepository.findById(Long.parseLong(userId)).orElseThrow(() -> new BaseException(ErrorCode.USER_NONFOUND_FAILED));
         Board board = new Board();
         board.setTitle(boardCreateDto.getTitle());
         board.setContent(boardCreateDto.getContent());
         board.setBoardType(boardCreateDto.getBoardType());
         board.setRgdt(LocalDateTime.now());
+        board.setUsers(users);
         Board  saveboard = boardRepository.save(board);
         return saveboard.getId();
     }
     @Transactional
     // jpa 엔티티 변경 감지 기능이 작동하려면 트랜잭션 안에서 엔티티를 변경해야함 (update 시 필요)
     // 읽기전용 => transactional 필요없음
-    public Integer updateBoard(BoradUpdateDto boradUpdateDto) {
+    public Integer updateBoard(BoradUpdateDto boradUpdateDto, String token) {
         Board board = boardRepository.findById(boradUpdateDto.getId()).orElseThrow(()->new BaseException(ErrorCode.BOARD_READ_FAILED));
+        String userId = jwtHTokenHeaderFilter.getUserIdFromToken(token);
+        Users users = userRepository.findById(Long.parseLong(userId)).orElseThrow(() -> new BaseException(ErrorCode.USER_NONFOUND_FAILED));
+        if (!board.getUsers().getId().equals(users.getId())) {
+            throw new BaseException(ErrorCode.BOARD_USER_UPDATE_FAILED);
+        }
+
         board.update(boradUpdateDto.getTitle(), boradUpdateDto.getContent());
         return board.getId();
     }
     @Transactional
-    public Integer deleteBoard(int id) {
+    public Integer deleteBoard(int id, String token) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new BaseException(ErrorCode.BOARD_DELETE_FAILED));
+        String userId = jwtHTokenHeaderFilter.getUserIdFromToken(token);
+        Users users = userRepository.findById(Long.parseLong(userId)).orElseThrow(() -> new BaseException(ErrorCode.USER_NONFOUND_FAILED));
+        if (!board.getUsers().getId().equals(users.getId())) {
+            throw new BaseException(ErrorCode.BOARD_DELETE_FAILED);
+        }
         boardRepository.delete(board);
         return board.getId();
     }
